requirements.txt
streamlit
google-generativeai
python-dotenv
fpdf2
imageio-ffmpeg
pandas
